+++ backend/controllers/ActividadRecienteController.php
@@ -55,7 +55,7 @@ class ActividadRecienteController extends Controller
     /**
      * Finds the ActividadReciente model based on its primary key value.
      * If the model is not found, a 404 HTTP exception will be thrown.
-     * @param integer $id
+     * @param int $id
      * @return ActividadReciente the loaded model
      * @throws NotFoundHttpException if the model cannot be found
      */
