+++ backend/controllers/PublicacionesController.php
@@ -4,11 +4,8 @@ namespace backend\controllers;
 
 use Yii;
 use yii\filters\AccessControl;
-
-use common\models\Publicaciones;
 use common\models\PublicacionesSearch;
 use yii\web\Controller;
-use yii\web\NotFoundHttpException;
 
 /**
  * PublicacionesController implements the CRUD actions for Publicaciones model.
@@ -16,6 +13,7 @@ use yii\web\NotFoundHttpException;
 class PublicacionesController extends Controller
 {
     use \common\utilities\Permisos;
+    use \common\traitrollers\PublicacionesComun;
     /**
      * @inheritdoc
      */
@@ -45,68 +43,5 @@ class PublicacionesController extends Controller
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
         ]);
-    }
-
-    /**
-     * Displays a single Publicaciones model.
-     * @param integer $id
-     * @return mixed
-     * @throws NotFoundHttpException if the model cannot be found
-     */
-    public function actionView($id)
-    {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
-    }
-
-    /**
-     * Updates an existing Publicaciones model.
-     * If update is successful, the browser will be redirected to the 'view' page.
-     * @param integer $id
-     * @return mixed
-     * @throws NotFoundHttpException if the model cannot be found
-     */
-    public function actionUpdate($id)
-    {
-        $model = $this->findModel($id);
-
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
-        }
-
-        return $this->render('update', [
-            'model' => $model,
-        ]);
-    }
-
-    /**
-     * Deletes an existing Publicaciones model.
-     * If deletion is successful, the browser will be redirected to the 'index' page.
-     * @param integer $id
-     * @return mixed
-     * @throws NotFoundHttpException if the model cannot be found
-     */
-    public function actionDelete($id)
-    {
-        $this->findModel($id)->delete();
-
-        return $this->redirect(['index']);
-    }
-
-    /**
-     * Finds the Publicaciones model based on its primary key value.
-     * If the model is not found, a 404 HTTP exception will be thrown.
-     * @param integer $id
-     * @return Publicaciones the loaded model
-     * @throws NotFoundHttpException if the model cannot be found
-     */
-    protected function findModel($id)
-    {
-        if (($model = Publicaciones::findOne($id)) !== null) {
-            return $model;
-        }
-
-        throw new NotFoundHttpException('The requested page does not exist.');
     }
 }
