+++ backend/controllers/UsuariosDatosController.php
@@ -33,7 +33,7 @@ class UsuariosDatosController extends Controller
     /**
      * Updates an existing UsuariosDatos model.
      * If update is successful, the browser will be redirected to the 'view' page.
-     * @param integer $id
+     * @param int $id
      * @return mixed
      * @throws NotFoundHttpException if the model cannot be found
      */
@@ -53,7 +53,7 @@ class UsuariosDatosController extends Controller
     /**
      * Finds the UsuariosDatos model based on its primary key value.
      * If the model is not found, a 404 HTTP exception will be thrown.
-     * @param integer $id
+     * @param int $id
      * @return UsuariosDatos the loaded model
      * @throws NotFoundHttpException if the model cannot be found
      */
