+++ common/traitrollers/TodoComun.php
@@ -1,0 +1,41 @@
+<?php
+
+namespace common\traitrollers;
+
+use Yii;
+
+use yii\web\NotFoundHttpException;
+
+/**
+ *
+ */
+trait TodoComun
+{
+
+    /**
+     * Deletes an existing Publicaciones model.
+     * If deletion is successful, the browser will be redirected to the 'index' page.
+     * @param int $id
+     * @return mixed
+     * @throws NotFoundHttpException if the model cannot be found
+     */
+    public function actionDelete($id)
+    {
+        $this->findModel($id)->delete();
+
+        return $this->redirect(['index']);
+    }
+
+    /**
+     * Displays a single Personajes model.
+     * @param int $id
+     * @return mixed
+     * @throws NotFoundHttpException if the model cannot be found
+     */
+    public function actionView($id)
+    {
+        return $this->render('view', [
+            'model' => $this->findModel($id),
+        ]);
+    }
+}
