+++ frontend/controllers/SeguidoresController.php
@@ -84,7 +84,7 @@ class SeguidoresController extends Controller
     /**
      * Finds the Seguidores model based on its primary key value.
      * If the model is not found, a 404 HTTP exception will be thrown.
-     * @param integer $id
+     * @param int $id
      * @return Seguidores the loaded model
      * @throws NotFoundHttpException if the model cannot be found
      */
